requirements.txt
openai>=1.0.0
python-docx>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0
